refactor: (리뷰반영) 댓글 삭제 기능 로직 서비스로 이동

+++ src/main/java/kr/codesqaud/cafe/comment/service/CommentService.java
@@ -4,6 +4,7 @@ import kr.codesqaud.cafe.comment.domain.Comment;
 import kr.codesqaud.cafe.comment.dto.RequestCommentForm;
 import kr.codesqaud.cafe.comment.mapper.CommentDtoMapper;
 import kr.codesqaud.cafe.comment.repository.CommentRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -25,7 +26,16 @@ public class CommentService {
         return commentRepository.findByArticleId(id);
     }
 
-    public void delete(long commentId) {
-        commentRepository.deleteById(commentId);
+    public HttpStatus delete(long commentId, String userId, String sessionId) {
+        try {
+            if (sessionId.equals(userId)) {
+                commentRepository.deleteById(commentId);
+                return HttpStatus.OK;
+            } else {
+                return HttpStatus.FORBIDDEN;
+            }
+        } catch (Exception e) {
+            return HttpStatus.INTERNAL_SERVER_ERROR;
+        }
     }
 }
